[#936] Extract text-family and dictType tags into tags/ (item 5)
Created `DisplayCAL/icc_profile/tags/text.py` with `TextDescriptionType`,
`TextType`, `MultiLocalizedUnicodeType`, `SignatureType`, and
`MakeAndModelType`, and `tags/dict_type.py` with `DictType` and
`DictTypeJSONEncoder`, moved out of `icc_profile/__init__.py`.

Also resolved the deferred `TextDescriptionType`/`MultiLocalizedUnicodeType`
import that `tags/colorant.py`'s `ProfileSequenceDescType` picked up in
item 4, now a direct top-level import since `tags/text.py` has no
dependency back on `tags/colorant.py`.
