+++ DisplayCAL/icc_profile/__init__.py
@@ -186,7 +186,15 @@ from DisplayCAL.icc_profile.tags.colorant import (
     ViewingConditionsType,
 )
 from DisplayCAL.icc_profile.tags.curve import CurveType, ParametricCurveType
+from DisplayCAL.icc_profile.tags.dict_type import DictType, DictTypeJSONEncoder
 from DisplayCAL.icc_profile.tags.lut import LUT16Type
+from DisplayCAL.icc_profile.tags.text import (
+    MakeAndModelType,
+    MultiLocalizedUnicodeType,
+    SignatureType,
+    TextDescriptionType,
+    TextType,
+)
 
 
 def create_RGB_A2B_XYZ(  # noqa: N802
@@ -2953,764 +2961,6 @@ def _mp_hdr_tonemap(
     elif its_hi:
         print("Max iterations", its_hi)
     return HDR_XYZ
-
-
-class DictType(ICCProfileTag, AODict):
-    """ICC dictType Tag.
-
-    Implements all features of 'Dictionary Type and Metadata TAG Definition'
-    (ICC spec revision 2010-02-25), including shared data (the latter will
-    only be effective for mutable types, ie. MultiLocalizedUnicodeType)
-
-    Examples:
-
-    tag[key]   Returns the (non-localized) value
-    tag.getname(key, locale='en_US') Returns the localized name if present
-    tag.getvalue(key, locale='en_US') Returns the localized value if present
-    tag[key] = value   Sets the (non-localized) value
-
-    """
-
-    def __init__(
-        self,
-        tagData: None | bytes = None,  # noqa: N803
-        tagSignature: None | str = None,  # noqa: N803
-    ) -> None:
-        ICCProfileTag.__init__(self, tagData, tagSignature)
-        AODict.__init__(self)
-        if not tagData:
-            return
-        numrecords = uInt32Number(tagData[8:12])
-        recordlen = uInt32Number(tagData[12:16])
-        if recordlen not in (16, 24, 32):
-            print(
-                f"Error (non-critical): '{tagData[:4]}' invalid record length "
-                f"(expected 16, 24 or 32, got {recordlen})"
-            )
-            return
-        elements = {}
-        for n in range(numrecords):
-            record = tagData[16 + n * recordlen : 16 + (n + 1) * recordlen]
-            if len(record) < recordlen:
-                print(
-                    f"Error (non-critical): '{tagData[:4]}' record {n} too short "
-                    f"(expected {recordlen} bytes, got {len(record)} bytes)"
-                )
-                break
-            for key, offsetpos in (
-                ("name", 0),
-                ("value", 8),
-                ("display_name", 16),
-                ("display_value", 24),
-            ):
-                if (
-                    offsetpos in (0, 8)
-                    or recordlen == offsetpos + 8
-                    or recordlen == offsetpos + 16
-                ):
-                    # Required:
-                    # Bytes 0..3, 4..7: Name offset and size
-                    # Bytes 8..11, 12..15: Value offset and size
-                    # Optional:
-                    # Bytes 16..23, 24..23: Display name offset and size
-                    # Bytes 24..27, 28..31: Display value offset and size
-                    offset = uInt32Number(record[offsetpos : offsetpos + 4])
-                    size = uInt32Number(record[offsetpos + 4 : offsetpos + 8])
-                    if offset > 0:
-                        if (offset, size) in elements:
-                            # Use existing element if same offset and size
-                            # This will really only make a difference for
-                            # mutable types i.e. MultiLocalizedUnicodeType
-                            data = elements[(offset, size)]
-                        else:
-                            data = tagData[offset : offset + size]
-                            try:
-                                if key.startswith("display_"):
-                                    data = MultiLocalizedUnicodeType(data, "mluc")
-                                else:
-                                    data = data.decode("UTF-16-BE", "replace").rstrip(
-                                        "\0"
-                                    )
-                            except Exception:
-                                print(
-                                    "Error (non-critical): could not decode "
-                                    f"'{tagData[:4]}', offset {offset}, length {size}"
-                                )
-                            # Remember element by offset and size
-                            elements[(offset, size)] = data
-                        if key == "name":
-                            name = data
-                            self[name] = ""
-                        else:
-                            self.get(name)[key] = data
-
-    def __getitem__(self, name: str) -> Any:  # noqa: ANN401
-        """Get item from dict.
-
-        Args:
-            name (str): Name of the item.
-
-        Returns:
-            Any: Value of the item.
-        """
-        return self.get(name).value
-
-    def __setitem__(self, name: str, value: Any) -> None:  # noqa: ANN401
-        """Set item in dict.
-
-        Args:
-            name (str): Name of the item.
-            value (Any): Value of the item.
-        """
-        AODict.__setitem__(self, name, ADict(value=value))
-
-    @property
-    def tagData(self) -> bytes:  # noqa: N802
-        """Return raw tag data.
-
-        Returns:
-            bytes: The raw tag data representing the dictionary.
-        """
-        numrecords = len(self)
-        recordlen = 16
-        keys = ("name", "value")
-        for value in self.values():
-            if not isinstance(value, dict):
-                continue
-            if "display_value" in value:
-                recordlen = 32
-                break
-            if "display_name" in value:
-                recordlen = 24
-        if recordlen > 16:
-            keys += ("display_name",)
-        if recordlen > 24:
-            keys += ("display_value",)
-        tag_data = [
-            b"dict",
-            b"\0" * 4,
-            uInt32Number_tohex(numrecords),
-            uInt32Number_tohex(recordlen),
-        ]
-        storage_offset = 16 + numrecords * recordlen
-        storage = []
-        elements = []
-        offsets = []
-        for item in self.items():
-            for key in keys:
-                if key == "name":
-                    element = item[0]
-                else:
-                    element = item[1].get(key) if isinstance(item[1], dict) else item[1]
-                if element is None:
-                    offset = 0
-                    size = 0
-                elif element in elements:
-                    # Use existing offset and size if same element
-                    offset, size = offsets[elements.index(element)]
-                else:
-                    offset = storage_offset + len(b"".join(storage))
-                    if isinstance(element, MultiLocalizedUnicodeType):
-                        data = element.tagData
-                    else:
-                        data = str(element).encode("UTF-16-BE")
-                    size = len(data)
-                    if isinstance(element, MultiLocalizedUnicodeType):
-                        # Remember element, offset and size
-                        elements.append(element)
-                        offsets.append((offset, size))
-                    # Pad all data with binary zeros so it lies on
-                    # 4-byte boundaries
-                    padding = math.ceil(size / 4.0) * 4 - size
-                    data += b"\0" * padding
-                    storage.append(data)
-                tag_data.append(uInt32Number_tohex(offset))
-                tag_data.append(uInt32Number_tohex(size))
-        tag_data.extend(storage)
-        return b"".join(tag_data)
-
-    @tagData.setter
-    def tagData(self, tagData: bytes) -> None:  # noqa: N802, N803
-        """Set raw tag data.
-
-        Does nothing, as the tagData is read-only.
-
-        Args:
-            tagData (bytes): The raw tag data to set.
-        """
-
-    def getname(
-        self,
-        name: str,
-        default: None | Any = None,  # noqa: ANN401
-        locale: str = "en_US",
-    ) -> str:
-        """Convenience function to get (localized) names.
-
-        Args:
-            name (str): The name of the item to get.
-            default (Any, optional): Default value to return if the item is not
-                found. Defaults to None.
-            locale (str, optional): Locale to use for localized names. Defaults
-                to "en_US".
-
-        Returns:
-            str: The localized name of the item if available, otherwise the
-                default value or the non-localized name.
-        """
-        item = self.get(name, default)
-        if item is default:
-            return default
-        if locale and "display_name" in item:
-            return item.display_name.get_localized_string(*locale.split("_"))
-        return name
-
-    def getvalue(
-        self,
-        name: str,
-        default: None | Any = None,  # noqa: ANN401
-        locale: str = "en_US",
-    ) -> Any:  # noqa: ANN401
-        """Convenience function to get (localized) values.
-
-        Args:
-            name (str): The name of the item to get.
-            default (Any, optional): Default value to return if the item is not
-                found. Defaults to None.
-            locale (str, optional): Locale to use for localized values.
-                Defaults to "en_US".
-
-        Returns:
-            Any: The localized value of the item if available, otherwise the
-                default value or the non-localized value.
-        """
-        item = self.get(name, default)
-        if item is default:
-            return default
-        if locale and "display_value" in item:
-            return item.display_value.get_localized_string(*locale.split("_"))
-        if isinstance(item, dict):
-            return item.value
-        return item
-
-    def setitem(
-        self,
-        name: str,
-        value: Any,  # noqa: ANN401
-        display_name: None | dict = None,
-        display_value: None | dict = None,
-    ) -> None:
-        """Convenience function to set items.
-
-        display_name and display_value (if given) should be dict types with
-        country -> language -> string mappings, e.g.:
-
-        {"en": {"US": u"localized string"},
-         "de": {"DE": u"localized string", "CH": u"localized string"}}
-
-
-        Args:
-            name (str): The name of the item to set.
-            value (Any): The value to set for the item.
-            display_name (None | dict, optional): Localized display names for
-                the item.
-            display_value (None | dict, optional): Localized display values
-                for the item.
-        """
-        self[name] = value
-        item = self.get(name)
-        if display_name:
-            item.display_name = MultiLocalizedUnicodeType()
-            item.display_name.update(display_name)
-        if display_value:
-            item.display_value = MultiLocalizedUnicodeType()
-            item.display_value.update(display_value)
-
-    def to_json(
-        self, encoding: str = "UTF-8", errors: str = "replace", locale: str = "en_US"
-    ) -> str:
-        """Return a JSON representation.
-
-        Display names/values are used if present.
-
-        Args:
-            encoding (str, optional): Encoding to use for the JSON string.
-                Defaults to "UTF-8".
-            errors (str, optional): Error handling scheme for encoding.
-                Defaults to "replace".
-            locale (str, optional): Locale to use for localized names/values.
-                Defaults to "en_US".
-
-        Returns:
-            str: JSON representation of the DictType object.
-        """
-        return DictTypeJSONEncoder(locale=locale).encode(self)
-
-
-class DictTypeJSONEncoder(json.JSONEncoder):
-    """JSON Encoder for the DictType class."""
-
-    def __init__(self, *args, **kwargs) -> None:
-        self.locale = kwargs.pop("locale") or "en_US"
-        super().__init__(*args, **kwargs)
-
-    def default(self, obj: Any) -> dict:  # noqa: ANN401
-        """Default method for encoding objects to JSON.
-
-        Args:
-            obj (object): The object to encode.
-
-        Returns:
-            dict: Encoded object as a dictionary.
-        """
-        return_data = {}
-        regex = re.compile(r"\\x([0-9a-f]{2})")
-        repl_str = r"\\u00\1"
-        for name in obj:
-            value = obj.getvalue(name, None, self.locale)
-            name = obj.getname(name, None, self.locale)
-            value = '"{}"'.format(repr(str(value))[2:-1].replace('"', '\\"'))
-            name = regex.sub(repl_str, name)
-            value = regex.sub(repl_str, value)
-            return_data[name] = value
-        return return_data
-
-
-class MakeAndModelType(ICCProfileTag, ADict):
-    """ICC makeAndModelType tag.
-
-    Args:
-        tagData (bytes): The raw tag data.
-        tagSignature (str): The signature of the tag.
-    """
-
-    def __init__(self, tagData: bytes, tagSignature: str) -> None:  # noqa: N803
-        ICCProfileTag.__init__(self, tagData, tagSignature)
-        self.update({"manufacturer": tagData[10:12], "model": tagData[14:16]})
-
-
-class MultiLocalizedUnicodeType(ICCProfileTag, AODict):  # ICC v4
-    """ICC v4 MultiLocalizedUnicodeType tag.
-
-    Args:
-        tagData (None | bytes, optional): Raw tag data. Defaults to None.
-        tagSignature (None | str, optional): Tag signature. Defaults to None.
-    """
-
-    def __init__(
-        self,
-        tagData: None | bytes = None,  # noqa: N803
-        tagSignature: None | str = None,  # noqa: N803
-    ) -> None:
-        ICCProfileTag.__init__(self, tagData, tagSignature)
-        AODict.__init__(self)
-        if not tagData:
-            return
-        records_count = uInt32Number(tagData[8:12])
-        record_size = uInt32Number(tagData[12:16])  # 12
-        if record_size != 12:
-            print(
-                f"Warning (non-critical): '{tagData[:4]}' invalid record length "
-                f"(expected 12, got {record_size})"
-            )
-            record_size = max(record_size, 12)
-        records = tagData[16 : 16 + record_size * records_count]
-        for _count in range(records_count):
-            record = records[:record_size]
-            if len(record) < 12:
-                continue
-            record_language_code = record[:2].decode("ascii", "replace")
-            record_country_code = record[2:4].decode("ascii", "replace")
-            record_length = uInt32Number(record[4:8])
-            record_offset = uInt32Number(record[8:12])
-            self.add_localized_string(
-                record_language_code,
-                record_country_code,
-                str(
-                    tagData[record_offset : record_offset + record_length],
-                    "utf-16-be",
-                    "replace",
-                ),
-            )
-            records = records[record_size:]
-
-    def __str__(self) -> str:
-        """Return tag as string.
-
-        Returns:
-            str: The first localized string in the tag, or an empty string if
-                no localized strings are available.
-        """
-        # TODO: Needs some work re locales
-        # (currently if en-UK or en-US is not found, simply the first entry
-        # is returned)
-        if "en" in self:
-            for country_code in ("UK", "US"):
-                if country_code in self["en"]:
-                    return self["en"][country_code]
-            if self["en"]:
-                # return first value
-                return next(iter(self["en"].values()))
-            return ""
-        if len(self):
-            # return first value of the first dictionary
-            return next(iter(next(iter(self.values())).values()))
-        return ""
-
-    def add_localized_string(
-        self, languagecode: str, countrycode: str, localized_string: str
-    ) -> None:
-        """Convenience function for adding localized strings."""
-        if languagecode not in self:
-            self[languagecode] = AODict()
-        self[languagecode][countrycode] = localized_string.strip("\0")
-
-    def get_localized_string(
-        self, languagecode: str = "en", countrycode: str = "US"
-    ) -> str:
-        """Convenience function for retrieving localized strings.
-
-        Falls back to first locale available if the requested one isn't
-
-        Args:
-            languagecode (str): The language code to retrieve the string for.
-                Defaults to "en".
-            countrycode (str): The country code to retrieve the string for.
-                Defaults to "US".
-
-        Returns:
-            str: The localized string for the given language and country code,
-                or the first available string if the requested one is not
-                found.
-        """
-        try:
-            return self[languagecode][countrycode]
-        except KeyError:
-            return str(self)
-
-    @property
-    def tagData(self) -> bytes:  # noqa: N802
-        """Return raw tag data."""
-        tag_data = [b"mluc", b"\0" * 4]
-        records_count = 0
-        for language_code in self:
-            for _ in self[language_code]:
-                records_count += 1
-        tag_data.append(uInt32Number_tohex(records_count))
-        record_size = 12
-        tag_data.append(uInt32Number_tohex(record_size))
-        storage_offset = 16 + record_size * records_count
-        storage = []
-        offsets = []
-        for language_code in self:
-            for country_code in self[language_code]:
-                tag_data.append((language_code + country_code).encode("ascii"))
-                data = self[language_code][country_code].encode("UTF-16-BE")
-                if data in storage:
-                    offset, record_length = offsets[storage.index(data)]
-                else:
-                    record_length = len(data)
-                    offset = len("".join(storage))
-                    offsets.append((offset, record_length))
-                    storage.append(data)
-                tag_data.append(uInt32Number_tohex(record_length))
-                tag_data.append(uInt32Number_tohex(storage_offset + offset))
-        tag_data.append(
-            b"".join(storage)
-        )  # TODO: Are you sure that this needs to be bytes
-        return b"".join(tag_data)
-
-    @tagData.setter
-    def tagData(self, tagData: bytes) -> None:  # noqa: N802, N803
-        """Set raw tag data.
-
-        Does nothing, as this tag is read-only.
-
-        Args:
-            tagData (bytes): Raw tag data to set.
-        """
-
-
-def SignatureType(tagData: bytes, tagSignature: str) -> Text:  # noqa: N802, N803
-    """Generate ICC signatureType tag.
-
-    Args:
-        tagData (bytes): The raw tag data containing the signature.
-        tagSignature (str): The signature of the tag.
-
-    Returns:
-        Text: An instance of the Text class representing the tag.
-    """
-    tag = Text(tagData[8:12].rstrip(b"\0"))
-    tag.tagData = tagData
-    tag.tagSignature = tagSignature
-    return tag
-
-
-class TextDescriptionType(ICCProfileTag, ADict):  # ICC v2
-    """ICC textDescriptionType tag.
-
-    Args:
-        tagData (None | bytes, optional): Raw tag data. Defaults to None.
-        tagSignature (None | str, optional): Tag signature. Defaults to None.
-    """
-
-    def __init__(
-        self,
-        tagData: None | bytes = None,  # noqa: N803
-        tagSignature: None | str = None,  # noqa: N803
-    ) -> None:
-        ICCProfileTag.__init__(self, tagData, tagSignature)
-        self.ASCII = b""
-        if not tagData:
-            return
-        ascii_description_length = uInt32Number(tagData[8:12])
-        if ascii_description_length:
-            ascii_description = tagData[12 : 12 + ascii_description_length].strip(
-                b"\0\n\r "
-            )
-            if ascii_description:
-                self.ASCII = ascii_description
-        unicode_offset = 12 + ascii_description_length
-        self.unicodeLanguageCode = uInt32Number(
-            tagData[unicode_offset : unicode_offset + 4]
-        )
-        unicode_description_length = uInt32Number(
-            tagData[unicode_offset + 4 : unicode_offset + 8]
-        )
-        if unicode_description_length:
-            if unicode_offset + 8 + unicode_description_length * 2 > len(tagData):
-                # Damn you MS. The Unicode character count should be the number of
-                # double-byte characters (including trailing unicode NUL), not the
-                # number of bytes as in the profiles created by Vista and later
-                print(
-                    f"Warning (non-critical): '{tagData[:4]}' Unicode part end points "
-                    "past the tag data, assuming number of bytes instead "
-                    "of number of characters for length"
-                )
-                unicode_description_length /= 2
-            if (
-                tagData[
-                    unicode_offset + 8 + unicode_description_length : unicode_offset
-                    + 8
-                    + unicode_description_length
-                    + 2
-                ]
-                == b"\0\0"
-            ):
-                print(
-                    f"Warning (non-critical): '{tagData[:4]}' Unicode part "
-                    "seems to be a single-byte string (double-byte "
-                    "string expected)"
-                )
-                char_bytes = 1  # fix for fubar'd desc
-            else:
-                char_bytes = 2
-            unicode_description = tagData[
-                unicode_offset + 8 : unicode_offset
-                + 8
-                + (unicode_description_length) * char_bytes
-            ]
-            try:
-                if char_bytes == 1:
-                    unicode_description = str(unicode_description, errors="replace")
-                elif unicode_description[:2] == b"\xfe\xff":
-                    # UTF-16 Big Endian
-                    if DEBUG:
-                        print("UTF-16 Big endian")
-                    unicode_description = unicode_description[2:]
-                    if (
-                        len(unicode_description.split(b" "))
-                        == unicode_description_length - 1
-                    ):
-                        print(
-                            f"Warning (non-critical): '{tagData[:4]}' "
-                            "Unicode part starts with UTF-16 big "
-                            "endian BOM, but actual contents seem "
-                            "to be UTF-16 little endian"
-                        )
-                        # fix fubar'd desc
-                        unicode_description = str(
-                            b"\0".join(unicode_description.split(b" ")),
-                            "utf-16-le",
-                            errors="replace",
-                        )
-                    else:
-                        unicode_description = str(
-                            unicode_description, "utf-16-be", errors="replace"
-                        )
-                elif unicode_description[:2] == b"\xff\xfe":
-                    # UTF-16 Little Endian
-                    if DEBUG:
-                        print("UTF-16 Little endian")
-                    unicode_description = unicode_description[2:]
-                    if unicode_description[0] == b"\0":
-                        print(
-                            f"Warning (non-critical): '{tagData[:4]}' "
-                            "Unicode part starts with UTF-16 "
-                            "little endian BOM, but actual "
-                            "contents seem to be UTF-16 big "
-                            "endian"
-                        )
-                        # fix fubar'd desc
-                        unicode_description = str(
-                            unicode_description, "utf-16-be", errors="replace"
-                        )
-                    else:
-                        unicode_description = str(
-                            unicode_description, "utf-16-le", errors="replace"
-                        )
-                else:
-                    if DEBUG:
-                        print("ASSUMED UTF-16 Big Endian")
-                    unicode_description = str(
-                        unicode_description, "utf-16-be", errors="replace"
-                    )
-                unicode_description = unicode_description.strip("\0\n\r ")
-                if unicode_description:
-                    if unicode_description.find("\0") < 0:
-                        self.Unicode = unicode_description
-                    else:
-                        print(
-                            "Error (non-critical): could not decode "
-                            f"'{tagData[:4]}' Unicode part - null byte(s) "
-                            "encountered"
-                        )
-            except UnicodeDecodeError:
-                print(
-                    "UnicodeDecodeError (non-critical): could not "
-                    f"decode '{tagData[:4]}' Unicode part"
-                )
-        else:
-            char_bytes = 1
-        mac_offset = unicode_offset + 8 + unicode_description_length * char_bytes
-        self.macScriptCode = 0
-        if len(tagData) > mac_offset + 2:
-            self.macScriptCode = uInt16Number(tagData[mac_offset : mac_offset + 2])
-            mac_description_length = ord(tagData[mac_offset + 2 : mac_offset + 3])
-            if mac_description_length:
-                try:
-                    mac_description = str(
-                        tagData[
-                            mac_offset + 3 : mac_offset + 3 + mac_description_length
-                        ],
-                        "mac-" + ENCODINGS["mac"][self.macScriptCode],
-                        errors="replace",
-                    ).strip("\0\n\r ")
-                    if mac_description:
-                        self.Macintosh = mac_description
-                except KeyError:
-                    print(
-                        f"KeyError (non-critical): could not decode '{tagData[:4]}' "
-                        f"Macintosh part (unsupported encoding {self.macScriptCode})"
-                    )
-                except LookupError:
-                    print(
-                        f"LookupError (non-critical): could not decode '{tagData[:4]}' "
-                        "Macintosh part (unsupported encoding "
-                        f"'{ENCODINGS['mac'][self.macScriptCode]}')"
-                    )
-                except UnicodeDecodeError:
-                    print(
-                        "UnicodeDecodeError (non-critical): could not decode "
-                        f"'{tagData[:4]}' Macintosh part"
-                    )
-
-    @property
-    def tagData(self) -> bytes:  # noqa: N802
-        """Return raw tag data.
-
-        Returns:
-            bytes: The raw tag data for the textDescriptionType tag.
-        """
-        tag_data = [
-            b"desc",
-            b"\0" * 4,
-            uInt32Number_tohex(len(self.ASCII) + 1),  # count of ASCII chars + 1
-            self.ASCII + b"\0",  # ASCII desc, \0 terminated
-            uInt32Number_tohex(self.get("unicodeLanguageCode", 0)),
-        ]
-        if "Unicode" in self:
-            tag_data.extend(
-                [
-                    # count of Unicode chars + 2 (UTF-16-BE BOM + trailing UTF-16 NUL,
-                    #                             1 char = 2 byte)
-                    uInt32Number_tohex(len(self.Unicode) + 2),
-                    b"\xfe\xff" + self.Unicode.encode("utf-16-be", "replace") + b"\0\0",
-                ]
-            )  # Unicode desc, \0\0 terminated
-        else:
-            tag_data.append(uInt32Number_tohex(0))  # Unicode desc length = 0
-        tag_data.append(uInt16Number_tohex(self.get("macScriptCode", 0)))
-        if "Macintosh" in self:
-            mac_description = self.Macintosh[:66]
-            tag_data.extend(
-                [
-                    uInt8Number_tohex(
-                        len(mac_description) + 1
-                    ),  # count of Macintosh chars + 1
-                    mac_description.encode(
-                        "mac-" + ENCODINGS["mac"][self.get("macScriptCode", 0)],
-                        "replace",
-                    )
-                    + (b"\0" * (67 - len(mac_description))),
-                ]
-            )
-        else:
-            tag_data.extend([b"\0", b"\0" * 67])  # Mac desc length = 0
-        return b"".join(tag_data)
-
-    @tagData.setter
-    def tagData(self, tagData: bytes) -> None:  # noqa: N802, N803
-        """Set tag data.
-
-        Does nothing, as this tag is read-only.
-
-        Args:
-            tagData (bytes): The raw tag data to set.
-        """
-
-    def __str__(self) -> str:
-        """Return tag as string.
-
-        Returns:
-            str: The localized string if available, otherwise the ASCII
-                representation of the tag.
-        """
-        if "Unicode" not in self and len(str(self.ASCII)) < 67:
-            # Do not use Macintosh description if ASCII length >= 67
-            localized_types = ("Macintosh", "ASCII")
-        else:
-            localized_types = ("Unicode", "ASCII")
-
-        for localized_type in localized_types:
-            if localized_type not in self:
-                continue
-            value = self[localized_type]
-            if not isinstance(value, str):
-                # Even ASCII description may contain non-ASCII chars, so
-                # assume system encoding and convert to unicode, replacing
-                # unknown chars
-                value = value.decode("utf-8", "replace")
-            return value
-        return None
-
-
-def TextType(tagData: bytes, tagSignature: str) -> Text:  # noqa: N802, N803
-    """Generate an ICC textType tag.
-
-    Args:
-        tagData (bytes): The raw tag data.
-        tagSignature (str): The signature of the tag, usually "text".
-
-    Returns:
-        Text: An instance of the Text class containing the tag data and
-            signature.
-    """
-    tag = Text(tagData[8:].rstrip(b"\0"))
-    tag.tagData = tagData
-    tag.tagSignature = tagSignature
-    return tag
 
 
 class VideoCardGammaType(ICCProfileTag, ADict):
